+++ CHANGELOG.md
@@ -9,6 +9,21 @@ This changelog follows the [Keep a Changelog](http://keepachangelog.com) format.
 
 ---
 
+## [1.6.0] - 2025-02-09
+
+### Added ###
+- [#28] Added code `Invalid` (-1).
+- [#28]{Qt C++11} Qt C++11 variant which uses `enum class` for the codes.
+
+### Changed ###
+- {Qt} `networkErrorToStatusCode()` now returns a `Code` instead of `int`.
+
+### Removed ###
+- Testing with Visual Studio 2013 because of outdated CMake.
+
+
+---
+
 
 ## [1.5.0] - 2022-01-03 ##
 
@@ -91,9 +106,12 @@ Initial (actually unversioned) release.
 ---
 
 
+[1.6.0]: https://github.com/j-ulrich/http-status-codes-cpp/compare/1.5.0...1.6.0
 [1.5.0]: https://github.com/j-ulrich/http-status-codes-cpp/releases/tag/1.5.0
 [1.4.0]: https://github.com/j-ulrich/http-status-codes-cpp/releases/tag/1.4.0
 [1.3.0]: https://github.com/j-ulrich/http-status-codes-cpp/releases/tag/1.3.0
 [1.2.0]: https://github.com/j-ulrich/http-status-codes-cpp/releases/tag/1.2.0
 [1.1.1]: https://github.com/j-ulrich/http-status-codes-cpp/releases/tag/1.1.1
 [1.1.0]: https://github.com/j-ulrich/http-status-codes-cpp/releases/tag/1.1.0
+
+[unreleased]: https://github.com/j-ulrich/http-status-codes-cpp/compare/1.6.0...develop